Make the AST readonly so it can be stored in cache
Memory can be saved by storing the full AST on the
ESP filesystem like LittleFS. In the previous version
the AST token content changed when running a rule. That
implied the need to constantly update the rules on the
FS. Writing to the FS is slow and would pose a risk for
the degredation of the flash when done frequently.

A new valstack is introduced to fix this. The token
values on the AST now point to a specific valstack
byte. The value position on the varstack is now
updated in the new valstack instead of in the AST
making the AST readonly.