Introduce TVALUE tokens besides the TVAR tokens
Instead of having each TVAR token pointing to a value
on the variable stack it now first points to a TVALUE
token. The TVALUE then points to the value on the
variable stack.

The TVALUE token is stored on the AST so it's easier
to work with just the AST object. The additional space
needed to store each additional TVALUE tokens is still
calculated for in advance.

The TVALUE token gives various advantages
- The name of a variable was stored multiple times.
When a variable name is long this wastes a lot of space
- Setting a variable value outside the AST is hard
This required you to search for the first call to the
variable and assign the value to it. However, to find
the first variable we first needed to parse the AST.
We can now look through the AST for the first TVALUE
token with the variable name we want to assign and all
involved TVAR tokens will be updated immediatly.
- Values needed to be synced between variable token
Each variable token with the same name needed to
contain the same value when it wasn't overwritten with
another one. This whole synchronisation is dropped
because each variable already points to the same TVALUE
pointers.

The only downside is the additional token needed of the
max 32 tokens we have space for.